requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.3.9
cffi==1.12.3
colorama==0.4.1
coverage==4.5.3
decorator==4.4.0
feedparser==5.2.1
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
jedi==0.13.3
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.3
parso==0.4.0
pickleshare==0.7.5
pluggy==0.11.0
prompt-toolkit==2.0.9
py==1.8.0
pycparser==2.19
Pygments==2.4.0
pytest==4.5.0
pytest-cov==2.7.1
python-dotenv==0.10.2
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.2
whitenoise==4.0
